modify the script so that it prints out the value for this new argument
Copy the `demo.R` script into your `docopt_demo` repo, and then edit it to add a new positional argument that is optional. Also modify the script so that it prints out the value for this new argument in addition to the entire `docopt` object and `docopt` object type. Commit this change and push it to your `docopt_demo` repository.